Travis CI nightly fix
It looks like Travis CI will not use saved environmental
variable when PR is raised from a forked repo, that means
the nightly build can not be done with PR from forked repo.

This will only enable nightly build, if the commit is not
from branch master, or if the commit is from a PR.

Note this PR itself is from forked repo, it will
pass if the fix is correct.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>