Use localstack:localstack emulator to test kinesis
The test for Kinesis implementations used to be done
manually and this is very error prune and out-of-sync
with code changes.

This commit utilize the localstack:localstack emulator
which is a container that emulates AWS services (not
just limited to Kinesis but also many other services).

This commit changes the test so that Kinesis code changes
are well covered.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>