Add GIFDataset support
Gif is kind of special compared to many other image format
in that GIF could have multiple frames. It is typically
rendered in Web browsers as an animation.

In TensorFlow, it is possible to process gif with decode_gif though
it will read the whole file into memory directly.

The different of GIFDataset is that GIFDataset only read one
frame at a time so it is more compat and might have a better
memory footprint.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>