Add linkstatic = True for cc_library
For some reason, bazel will create two copies of libraries
(one in .a and one in .so) with cc_library if `linkstatic = True`
is not specified. That created duplicated .so files that are not
needed. This fix adds `linkstatic = True` to cc_library
so that no unneeded .so are distributed in whl files.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>