Expose tfio.IOTensor class and from_audio and tfio.IOTensor.to_dataset() (#437)
* Expose tfio.IOTensor class and from_audio and tfio.IOTensor.to_dataset()

This PR tries to expose a tfio.IOTensor which could be applied to and io related data which are indexable (__getitem__ and __len__)

The idea is to bind __getitem__ and __len__ to kernel ops in run time, so that is is not necessarily to read everything in memory.

The first file format is the WAV file. With tfio.IOTensor dtype and shape are exposed with __getitem__ and __len__.

Further, a rate property has been exposed specifically for Audio/WAV file which gives sample rate.

This tfio.IOTensor only works in eager mode.

In additional this PR also converts WavDataset to use IOTensor (instead of direct C++ implementation).

This PR also carries 420.

Note as was discussed, rebatch has been dropped. Instead, a PR to core tensorflow repo will be opened.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Remove Iterable from reference

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Pylint fix

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add a decorator so that it could be picked up by __repr__ automatically

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Fix python 3 issue

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add KafkaDataset to tensorflow_io.core.python.ops.kafka_ops.KafkaDataset

intend to deprecate old KafkaDataset soon.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add KafkaIOTensor which stores data in memory (so that it is indexable)

This is build around the same code base as KafkaDataset C++.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Deprecate WAVDataset, and pylint fix

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Remove leftover print

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Import GetTensorFlowType and GetArrowType

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Fix kokoro version

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>