Add tfio.IOTensor.from_prometheus support (#438)
* Add tfio.IOTensor.from_prometheus support

In the past couple of months I have been trying to come
up with a good example of using Prometheus with TensorFlow
for infrastructure/compute usage prediction and alerting
in case of anomaly.

My plan was to use LSTM to combine with prometheus for that.
Though there is a PrometheusDataset support, the format
of dataset is not very intuitive and hard to do it in
a smooth way.

The biggest challenge is that, normally for time series data
you have a look back window to train, and you normalize
the data (with total). Both are not easily available.
Further more, to process data in normal TF operations
such as tf.roll, it requires reading the whole data into
one Tenor. But this is not very straightforward as well with
iterable dataset.

This PR adds tfio.IOTensor.from_prometheus which
allows easily convertion the prometheus observation
data into a tuple of Tensors, and do additional feature
engineering.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Rename libtensorflow_io_prometheus.so to libtensorflow_io_golang.so

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Read single column at a time so that resource could be reused.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Change to shapes, dtypes to match TensorSpec order

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add `records` mode for `from_json` that matches `records` in pandas

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>